bug fixed at the target_policy_achieved
bug fixed at the target_policy_achieved
