+++ Environments/taxi_environment_wrapper.py
@@ -17,8 +17,12 @@ class TaxiSimpleEnv(TaxiEnv):
         super().__init__(num_taxis=1, num_passengers=1, max_fuel=None, domain_map=NEW_MAP,
                          collision_sensitive_domain=False)
 
-    def encode(self, taxi_row, taxi_col, fuel, pass_loc_x, pass_loc_y, dest_idx_x, dest_idx_y, pass_status):
+    def encode(self, state):
         # (self.num_rows), self.num_columns, max_fuel[0] + 1, self.num_rows, self.num_columns, self.passengers_locations, 4
+        taxi_row, taxi_col, fuel, pass_loc_x, pass_loc_y, dest_idx_x, dest_idx_y, pass_status = state[0], state[1], \
+                                                                                                state[2], state[3], \
+                                                                                                state[4], state[5], \
+                                                                                                state[6], state[7]
         dest_idx = self._get_pass_dest_idx(dest_idx_x, dest_idx_y)
 
         i = taxi_row
@@ -79,7 +83,7 @@ class TaxiSimpleEnv(TaxiEnv):
     def _get_pass_dest_idx(self, dest_idx_x, dest_idx_y):
         dest_idx = -1
         for i, loc in enumerate(self.passengers_locations):
-            if dest_idx_x == loc[0] and dest_idx_y == loc[1]:
+            if (dest_idx_x == loc[0] and dest_idx_y == loc[1]) or (dest_idx_x == 1 and dest_idx_y == 2):
                 dest_idx = i
                 break
         if dest_idx == -1:
@@ -146,7 +150,7 @@ class TaxiTransformedEnv(TaxiSimpleExampleEnv):
 
     def get_states_from_partial_obs(self, partial_obs):  # TODO Mira: add the multi agent case
         partial_obs_aligned_with_env = False
-        iter_num = 100
+        iter_num = 200
         while not partial_obs_aligned_with_env and iter_num != 0:
             obs = self.reset()
             obs = obs[list(obs.keys())[0]].tolist()[0]  # get the observation as a list
@@ -154,27 +158,30 @@ class TaxiTransformedEnv(TaxiSimpleExampleEnv):
                 partial_obs_aligned_with_env = True
             iter_num -= 1
 
-        taxi_x = [partial_obs[0]] if (partial_obs[0] is not None) else list(range(self.num_columns))
-        taxi_y = [partial_obs[1]] if (partial_obs[1] is not None) else list(range(self.num_rows))
-        fuel = [partial_obs[2]] if partial_obs[2] else list(range(self.max_fuel[0]))
-        passenger_start_x, passenger_start_y = [obs[3]], [obs[4]]
-        passenger_dest_x, passenger_dest_y = [obs[5]], [obs[6]]
-        passenger_status = [partial_obs[7]] if partial_obs[7] else list(range(1, 4))
-        states = list(
-            product(taxi_x, taxi_y, fuel, passenger_start_x, passenger_start_y, passenger_dest_x, passenger_dest_y,
-                    passenger_status, repeat=1))
-        states = [list(state) for state in states]
+        if partial_obs_aligned_with_env:
+            taxi_x = [partial_obs[0]] if (partial_obs[0] is not None) else list(range(self.num_columns))
+            taxi_y = [partial_obs[1]] if (partial_obs[1] is not None) else list(range(self.num_rows))
+            fuel = [partial_obs[2]] if partial_obs[2] else list(range(self.max_fuel[0]))
+            passenger_start_x, passenger_start_y = [obs[3]], [obs[4]]
+            passenger_dest_x, passenger_dest_y = [obs[5]], [obs[6]]
+            passenger_status = [partial_obs[7]] if partial_obs[7] else list(range(1, 4))
+            states = list(
+                product(taxi_x, taxi_y, fuel, passenger_start_x, passenger_start_y, passenger_dest_x, passenger_dest_y,
+                        passenger_status, repeat=1))
+            states = [list(state) for state in states]
+        else:
+            states = []
         return states
 
     def _is_aligned(self, obs, partial_obs):
         taxi_x, taxi_y = partial_obs[0], partial_obs[1]
         passenger_start_x, passenger_start_y, passenger_dest_x, passenger_dest_y = self._get_passenger_info(partial_obs)
-        return (taxi_x is None or passenger_start_x == obs[0]) and (
-                taxi_y is None or passenger_start_x == obs[1]) and (
-                passenger_start_x is None or passenger_start_x == obs[3]) and (
-                passenger_start_y is None or passenger_start_y == obs[4]) and (
-                passenger_dest_x is None or passenger_dest_x == obs[5]) and (
-                passenger_dest_y is None or passenger_dest_y == obs[6])
+        return (taxi_x is None or taxi_x == obs[0]) and (
+                taxi_y is None or taxi_y == obs[1]) and (
+                       passenger_start_x is None or passenger_start_x == obs[3]) and (
+                       passenger_start_y is None or passenger_start_y == obs[4]) and (
+                       passenger_dest_x is None or passenger_dest_x == obs[5]) and (
+                       passenger_dest_y is None or passenger_dest_y == obs[6])
 
     def _get_passenger_info(self, partial_obs):
         passenger_start_x, passenger_start_y = partial_obs[3], partial_obs[4]
