+++ utils.py
@@ -53,7 +53,7 @@ def target_policy_achieved(env, agent, target_policy):  # TODO Mira - to add the
         states_from_partial_obs = env().get_states_from_partial_obs(partial_obs)
         for i, state in enumerate(states_from_partial_obs):
             state = np.reshape(np.array(state), (1, len(state)))
-            action = agent.compute_action(state)
+            action = agent.compute_action(state, predict=True)
             if action != target_policy[original_partial_obs]:
                 num_of_failed_policies += 1
                 result = False
@@ -62,7 +62,8 @@ def target_policy_achieved(env, agent, target_policy):  # TODO Mira - to add the
     print(" ============================================================== ")
     print("=========> num_of_success_policies: ", num_of_success_policies)
     print("=========> num_of_failed_policies: ", num_of_failed_policies)
-    success_rate = num_of_success_policies / (num_of_success_policies + num_of_failed_policies)
+    all_policies = num_of_success_policies + num_of_failed_policies
+    success_rate = num_of_success_policies / all_policies if all_policies != 0 else 1
     print("=========> Success rate:", success_rate)
     print(" ============================================================== ")
     return success_rate > 0.8
